+++ src/neofoam/framework/context.py
@@ -32,4 +32,4 @@ class Context(BaseModel):
     fields: dict[str, Any]
     models: dict[str, Any]
     mesh: Any = None
-    runTime: Any = None
+    runtime: Any = None