+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - add openfoam reader and writer [#242](https://github.com/exasim-project/NeoFOAM/pull/242)
 - added neoicofoam [243](https://github.com/exasim-project/NeoFOAM/pull/243)
 - Allow auto grabbing version from submodule without initialization and update the documentation [#210](https://github.com/exasim-project/NeoFOAM/pull/210)
+- Refactor graph and init package  [#284](https://github.com/exasim-project/NeoFOAM/pull/284)
 
 ## Fixes
 - Fix spurious bad_any_cast errors when reading fixedValue boundaries [#194](https://github.com/exasim-project/NeoFOAM/pull/194)